Q2/step3/requirements.txt
# 问题3：综合多因素的NIPT最佳时点优化
# Python依赖包列表

# 基础科学计算库
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# 机器学习库
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0

# 统计分析库
lifelines>=0.27.0
statsmodels>=0.13.0

# 数据处理和I/O
openpyxl>=3.0.0
PyYAML>=5.4.0
joblib>=1.1.0

# 可视化库（可选）
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# 数值优化
cvxpy>=1.2.0

# 进度条和工具
tqdm>=4.62.0
pathlib2>=2.3.6

# 测试库（可选）
pytest>=6.2.0
pytest-cov>=3.0.0

# Jupyter支持（可选）
jupyter>=1.0.0
ipython>=7.0.0
notebook>=6.4.0

# 性能分析工具（可选）
memory-profiler>=0.60.0
line-profiler>=3.5.0